refactor(updaters): change db reducer into updater
+ impact rootReducer temporarly

Ref #4
