+++ src/store/modules/reducer.js
@@ -3,7 +3,7 @@ const { combineReducers } = require('redux');
 const { withDefaultState } = require('redux-fp');
 const { toReducer } = require('redux-fun')
 
-const db = require('./db/reducer');
+const db = require('./db/updater');
 const logs = require('./logs/updater');
 
 const credentials = withDefaultState({}, always(identity))
@@ -11,7 +11,7 @@ const serverConfiguration = withDefaultState({}, always(identity))
 
 module.exports = combineReducers({
   logs: toReducer(logs),
-  db,
+  db: toReducer(db),
   credentials: toReducer(credentials),
   serverConfiguration: toReducer(serverConfiguration),
 });
